refactor(rpc): move RpcRouter to dedicated module
Extract RpcRouter into `router.rs` and re-export from `lib.rs`.
